+++ src/shared/card/delete.rs
@@ -2,15 +2,13 @@ use anyhow::Result;
 use clap::Parser;
 use pimalaya_cli::printer::{Message, Printer};
 
-use crate::shared::client::AddressbookClient;
+use crate::shared::{arg::AddressbookIdArg, client::AddressbookClient};
 
 /// Permanently delete the given card.
 #[derive(Debug, Parser)]
 pub struct CardDeleteCommand {
-    /// Identifier of the parent addressbook. Falls back to the
-    /// `addressbook.default` config when omitted.
-    #[arg(short = 'k', long = "addressbook", value_name = "ADDRESSBOOK-ID")]
-    pub addressbook_id: Option<String>,
+    #[command(flatten)]
+    pub addressbook: AddressbookIdArg,
     /// Card UID.
     #[arg(value_name = "CARD-ID")]
     pub card_id: String,
@@ -18,7 +16,7 @@ pub struct CardDeleteCommand {
 
 impl CardDeleteCommand {
     pub fn execute(self, printer: &mut impl Printer, mut client: AddressbookClient) -> Result<()> {
-        let addressbook_id = client.account.addressbook_id(self.addressbook_id)?;
+        let addressbook_id = client.account.addressbook_id(self.addressbook.id)?;
         client.delete_card(&addressbook_id, &self.card_id)?;
 
         let msg = format!("Card `{}` successfully deleted", self.card_id);