+++ src/shared/mod.rs
@@ -1,3 +1,4 @@
 pub mod addressbook;
+pub mod arg;
 pub mod card;
 pub mod client;